flask_app/requirements.txt

flask
psycopg2-binary